create functioning knightMoves functioning
the function uses breath first traversing for the graph and so far
works as intended. using the queue data structure im able to look for
the coordinate im trying to find easier because on each vertex i go
to i check if im there then return back. I wont be continuing any futher
because the next vertices are either the same or longer.