adjust what knightMoves returns
before it would return the first route found since it being breath first
traversal the first one found would the or one of the correct answers.
now the function returns and array of all the fastest routes
inside an array. I can do this because i can stop the program from
adding to the queue once a solution is found and therefore i will get
the remaining solutions that will all be the same amount of steps